% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrateBatch_inter.R
\name{calibrateBatch.inter.rlm}
\alias{calibrateBatch.inter.rlm}
\title{Inter batch calibration by robust linear modelling (rlm)}
\usage{
calibrateBatch.inter.rlm(
  data = ...,
  feature = feature,
  batch = batch,
  intensity = intensity
)
}
\arguments{
\item{data}{metabolomics in tidy long-format}

\item{feature}{column name of feature (by default feature)}

\item{batch}{column name of batch (by default batch)}

\item{intensity}{column name of intensity (by default intensity)}
}
\description{
Inter batch calibration by robust linear modelling (rlm)
}
