% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrateBatch_intra.R
\name{calibrateBatch.intra.rlm.old}
\alias{calibrateBatch.intra.rlm.old}
\title{Deprecated calibration method (without the support of custom column name)}
\usage{
calibrateBatch.intra.rlm.old(
  data = ...,
  intensity = intensity,
  injection_sequence = injection_sequence
)
}
\description{
Deprecated calibration method (without the support of custom column name)
}
