% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrateBatch_intra.R
\name{calibrateBatch.intra.rlm}
\alias{calibrateBatch.intra.rlm}
\title{Intra batch calibration by robust linear modelling (rlm)}
\usage{
calibrateBatch.intra.rlm(
  data = ...,
  intensity = intensity,
  injection_sequence = injection_sequence
)
}
\arguments{
\item{data}{Metabolomics data in long-format}

\item{intensity}{The column name of intensity (by default intensity)}

\item{injection_sequence}{The column name of injection sequence (by default injection_sequence)}
}
\description{
Intra batch calibration by robust linear modelling (rlm)
}
