% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeNA.R
\name{detect.birds}
\alias{detect.birds}
\title{Detect early and late eluting compounds (early- and late- birds)}
\usage{
detect.birds(rt = ..., limit_low = 0.3, limit_high = 6.5)
}
\arguments{
\item{rt}{retention time in the numeric format}

\item{limit_low}{lower limit of features to be removed}

\item{limit_high}{higher limit of features to be removed}
}
\value{
\code{rt} between \code{limit_low} and \code{limit-high}
}
\description{
Detect early and late eluting compounds (early- and late- birds)
}
