% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeNA.R
\name{imputeNA.old}
\alias{imputeNA.old}
\title{Impute missing data of metabolomics experiment}
\usage{
imputeNA.old(data = ...)
}
\arguments{
\item{data}{Tidy format dataframe consisting cols intensity and feature}
}
\description{
Impute missing data of metabolomics experiment
}
