% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeNA.R
\name{imputeNA}
\alias{imputeNA}
\title{Impute missing data of metabolomics experiment}
\usage{
imputeNA(data = ..., feature = feature, intensity = intensity)
}
\arguments{
\item{data}{Tidy format dataframe consisting cols intensity and feature}

\item{feature}{The column name of feature (by default feature)}

\item{intensity}{The column name of intensity (by default intensity)}
}
\description{
Impute missing data of metabolomics experiment
}
