% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFeature.R
\name{plotFeature}
\alias{plotFeature}
\title{Plot features by injection sequence}
\usage{
plotFeature(data = ..., Feature = ..., intensity = data$intensity)
}
\arguments{
\item{data}{a dataframe consisting intensity, feature and injection sequence}
}
\description{
Plot features by injection sequence
}
