% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobulk_de.R
\name{pseudobulk_de}
\alias{pseudobulk_de}
\title{Run pseudobulk differential expression methods}
\usage{
pseudobulk_de(
  input,
  meta = NULL,
  replicate_col = "replicate",
  cell_type_col = "cell_type",
  label_col = "label",
  min_cells = 3,
  min_reps = 2,
  min_features = 0,
  de_family = "pseudobulk",
  de_method = "edgeR",
  de_type = "LRT"
)
}
\arguments{
\item{input}{a single-cell matrix to be converted, with features (genes) in rows
and cells in columns. Alternatively, a \code{Seurat}, \code{monocle3}, or 
or \code{SingleCellExperiment} object can be directly input.}

\item{meta}{the accompanying meta data whereby the rownames match the column
names of \code{input}.}

\item{replicate_col}{the vector in \code{meta} containing the replicate 
information. Defaults to \code{replicate}.}

\item{cell_type_col}{the vector in \code{meta} containing the cell type 
information. Defaults to \code{cell_type}.}

\item{label_col}{the vector in \code{meta} containing the experimental
label. Defaults to \code{label}.}

\item{min_cells}{the minimum number of cells in a cell type to retain it.
Defaults to \code{3}.}

\item{min_reps}{the minimum number of replicates in a cell type to retain it.
Defaults to \code{2}.}

\item{min_features}{the minimum number of counts for a gene to retain it.
Defaults to \code{0}}

\item{de_method}{the specific differential expression testing method to use.
Defaults to edgeR.}

\item{de_type}{the specific parameter of the differential expression testing
method. Defaults to LRT for edgeR, LRT for DESeq2, and trend for limma.}
}
\value{
a data frame containing differential expression results.
}
\description{
Run pseudobulk differential expression methods on single-cell data
}
