% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{pull.data}
\alias{pull.data}
\title{Pull pre-processed data and summary statistics based on the regular expression (naming pattern)}
\usage{
\method{pull}{data}(pattern = ..., data_table. = data_table, sample_table. = sample_table)
}
\arguments{
\item{pattern}{regular expression in sample_table}

\item{data_table.}{data table}

\item{sample_table.}{sample information table}
}
\description{
Pull pre-processed data and summary statistics based on the regular expression (naming pattern)
}
