% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecell_de.R
\name{singlecell_de}
\alias{singlecell_de}
\title{Run single-cell differential expression}
\usage{
singlecell_de(
  input,
  meta = NULL,
  cell_type_col = "cell_type",
  label_col = "label",
  de_method = "wilcox",
  min_cells = 3,
  min_features = 0
)
}
\arguments{
\item{input}{a single-cell matrix to be converted, with features (genes) in rows
and cells in columns. Alternatively, a \code{Seurat}, \code{monocle3}, or 
or \code{SingleCellExperiment} object can be directly input.}

\item{meta}{the accompanying meta data whereby the rownames match the column
names of \code{input}.}

\item{cell_type_col}{the vector in \code{meta} containing the cell type 
information. Defaults to \code{cell_type}.}

\item{label_col}{the vector in \code{meta} containing the experimental
label. Defaults to \code{label}.}

\item{de_method}{the mixed model type to use. Defaults to wilcox.}

\item{min_cells}{the minimum number of cells in a cell type to retain it.
Defaults to \code{3}.}

\item{min_features}{the minimum number of counts for a gene to retain it.
Defaults to \code{0}}
}
\value{
a data frame containing differential expression results.
}
\description{
Run differential expression using traditional single-cell methods
Note that this is effectively a wrapper of the FindMarkers function in Seurat.
}
