\name{west10}
\alias{west10}
\title{Journey to the West, one of the Four Great Classical Novels of Chinese Literature by WU, Cheng'en.}
\description{
 The \code{west10} data frame has 408 rows and 10 columns. It records the appearance of top 10 characters in 408 scenes, the characters who appeared ("1") or not ("0") in each of the scenes. The dataset was collected via crowdsourcing in the classes of Mathematical Introduction to Data Analysis and Statistical Learning, taught by Prof. Yuan YAO at Peking University. he original data contains all the 302 characters, which can be downloaded at the sourse website.
}
\format{
  This file contains the following data frame:
  \describe{
    \item{west10}{a data frame of 408 rows and 10 columns}
  }
}
\details{
It records the appearance of top 10 characters in 408 scenes, the characters who appeared ("1") or not ("0") in each of the scenes. All the character names are in Chinese Pinyin.
}
%\source{
%  \url{http://math.stanford.edu/~yuany/course/2014.fall/}
%}
\references{
Yuan YAO, A Mathematical Introduction to Data Analysis, Lecture Notes in \emph{Peking University}
}
\keyword{datasets}
