% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.lb.R
\name{plot.lb}
\alias{plot.lb}
\title{Plot method for lb objects}
\usage{
\method{plot}{lb}(x, omit.zeros = TRUE, eps = 1e-10, ...)
}
\arguments{
\item{x}{lb object}

\item{omit.zeros}{When the number of variables  is much greater than
the number of observations, many coefficients will never be nonzero;
this logical (default \code{TRUE}) avoids plotting these zero coefficents}

\item{eps}{Definition of zero above, default is \code{1e-10}}

\item{\dots}{Additonal arguments for generic plot. Can be used to set xlims, change colors, line widths, etc}
}
\description{
Produce a plot of an LB fit. The default is a complete coefficient path.
}
\details{
The default plot uses the fraction of L1 norm as the x.
For multinomial case, the sum of absolute values of different class's
coefficients are caculated to represent each variable.
The intercept term is not ploted
}
\author{
Feng Ruan, Jiechao Xiong and Yuan Yao
}
\keyword{methods}

